requirements.txt
Django
mysql-connector-python
django-chartJS
pillow
babel
zoomus
twilio
channels
django-phonenumber-field[phonenumbers]
phonenumbers
Celery==4.4.7
flower
PyJWT==2.0.1
django-tz-detect
django_otp
django-guardian
python-magic
python-magic-bin
django-formtools


#to use run
#python -m pip install -r requirements.txt

###### Requirements without Version Specifiers ######`
#nose
#tesr45


###### Requirements with Version Specifiers ######`
#docopt == 0.6.1             # Version Matching. Must be version 0.6.1
#keyring >= 4.1.1            # Minimum version 4.1.1
#coverage != 3.5             # Version Exclusion. Anything except version 3.5
#Mopidy-Dirble ~= 1.1        # Compatible release. Same as >= 1.1, == 1.*
